backend/requirements.txt
mysqlclient==2.0.3
dateparser==1.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.6
django-3-jet==1.0.8
django-cors-headers==3.7.0
#django-heroku==0.3.1
django-simple-history==2.12.0
djangorestframework==3.12.2
drf-yasg==1.20.0
Faker==6.6.2
greenlet==1.0.0
gunicorn==20.0.4
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
twilio==6.56.0
SQLAlchemy==1.4.6
Pillow==8.2.0
image-quality==1.2.7
django-import-export==2.5.0
django-environ==0.4.5
jet-django==1.1.1
